+++ RELEASENOTES.md
@@ -15,7 +15,7 @@ All
 MBSim
 -----
 - Use ellipse parameters to define tyre contour instead of rim radius. For backward compatibiltiy rim radius is still supported by MBSimXML and MBSimGUI but it should no longer be used.
-
+- Enable feature "tyre side" and "contact point transformation" in implementation of Magic Formula 6.2
 
 MBSimGUI
 --------